Update PreprocessSchemas with interface implementation
