+++ preprocessor/src/main/java/dev/cdevents/package-info.java
@@ -14,4 +14,4 @@
  *
  */
 
-package dev.cdevents.preprocessor;
+package dev.cdevents;
