+++ preprocessor/src/main/java/dev/cdevents/process/ProcessSchemas.java
@@ -1,0 +1,10 @@
+package dev.cdevents.process;
+
+public interface ProcessSchemas {
+
+    /**
+     * Updates schemas in the specified input schemasPath.
+     * @param schemasPath
+     */
+    void updateSchemas(String schemasPath);
+}
